Update the implementation of PAGImageView. (#887)
* Switch PAGImageView cache to native implementation

* Update the implementation of PAGImageView on the iOS platform.

* Add the missing getComposition method.

* Update the url style of sandbox path.

* Fix BGRA texture support on swiftshader.

* Add numFrames in PAGImageView.

* Switch PAGImageView cache to native implementation

---------

+++ src/platform/ios/private/PAGDecoderImpl.mm
@@ -73,7 +73,7 @@
 }
 
 - (BOOL)copyFrameTo:(void*)pixels rowBytes:(size_t)rowBytes at:(NSInteger)index {
-  return pagDecoder->readFrame(index, pixels, rowBytes, pag::ColorType::BGRA_8888,
+  return pagDecoder->readFrame(static_cast<int>(index), pixels, rowBytes, pag::ColorType::BGRA_8888,
                                pag::AlphaType::Premultiplied);
 }
 
@@ -86,8 +86,8 @@
   CVPixelBufferLockBaseAddress(pixelBuffer, 0);
   auto rowBytes = CVPixelBufferGetBytesPerRowOfPlane(pixelBuffer, 0);
   auto pixels = CVPixelBufferGetBaseAddress(pixelBuffer);
-  auto success = pagDecoder->readFrame(index, pixels, rowBytes, pag::ColorType::BGRA_8888,
-                                       pag::AlphaType::Premultiplied);
+  auto success = pagDecoder->readFrame(static_cast<int>(index), pixels, rowBytes,
+                                       pag::ColorType::BGRA_8888, pag::AlphaType::Premultiplied);
   CVPixelBufferUnlockBaseAddress(pixelBuffer, 0);
   if (!success) {
     return nil;
